VSL 138: Refactor Safe Settings Page (#75)
* refactor structure done

* front-end refactor done

* merge with main

* fix home screen updating bug

* clean up

* remove redundant web3 context use
